Add files via upload
This project uses the Breast Cancer dataset from the UCI repository to classify whether a tumor is malignant or benign. We apply a Decision Tree algorithm to predict the diagnosis based on various features such as radius, texture, perimeter, area, and smoothness of the tumor cells. The goal is to build a model that accurately distinguishes between malignant and benign tumors.