Add files via upload
This project uses the UCI Online Shopper dataset to predict whether a visitor will make a purchase on an e-commerce website. The model leverages features such as time spent on site, product page views, and traffic sources to classify whether a visitor is likely to convert into a customer. The goal is to build a machine learning model to enhance customer targeting and improve conversion rates.